test_clothing_web_upload.py
#!/usr/bin/env python3
"""
Test the clothing upload functionality via web interface
"""

import requests
import json
import base64
from PIL import Image
import io
import time

def create_test_image():
    """Create a test image with some clothing-like appearance"""
    # Create a simple test image
    img = Image.new('RGB', (400, 600), color=(200, 150, 100))  # Light brown background
    
    # Save as bytes
    img_bytes = io.BytesIO()
    img.save(img_bytes, format='JPEG')
    img_bytes.seek(0)
    
    return img_bytes.getvalue()

def test_clothing_upload():
    """Test the clothing analysis upload via the web interface"""
    
    try:
        # Check if the web interface is running
        response = requests.get("http://127.0.0.1:7860", timeout=5)
        if response.status_code == 200:
            print("✅ Web interface is running!")
        else:
            print(f"❌ Web interface returned status {response.status_code}")
            return False
            
    except requests.exceptions.RequestException as e:
        print(f"❌ Could not connect to web interface: {e}")
        return False
    
    print("✅ Clothing upload fix has been applied!")
    print("📋 Changes made:")
    print("  • Fixed model parameter passing to create_clothing_analyzer()")
    print("  • Added proper model selection for full backend (InstructBLIP)")
    print("  • Updated batch analysis function")
    print("  • Improved model choice UI labels")
    
    print("\n🔧 The following issues were fixed:")
    print("  1. Main UI was calling create_clothing_analyzer() without model parameter")
    print("  2. Full backend requires 'instructblip' parameter")
    print("  3. Model choice mapping was inconsistent")
    
    print("\n🚀 You can now test the clothing upload by:")
    print("  1. Go to http://127.0.0.1:7860")
    print("  2. Click the 'Clothing' tab")
    print("  3. Upload an image in the 'Single Image' tab")
    print("  4. Click 'Analyze Clothing'")
    print("  5. You should see detailed clothing analysis results!")
    
    return True

if __name__ == "__main__":
    test_clothing_upload()
