requirements.txt

# UI and visualization
# gradio>=4.0.0
gradio #latest stable version

# PyTorch 2.7.0 with CUDA 12.8 (Modern Blackwell/RTX 5000 series support)
# Using official PyTorch wheels (A1111 approach) instead of Early Access
torch==2.7.0
torchvision==0.22.0
torchaudio==2.7.0
--extra-index-url https://download.pytorch.org/whl/cu128

# Legacy Early Access wheels (kept for reference, uncomment if needed)
# https://huggingface.co/w-e-w/torch-2.6.0-cu128.nv/resolve/main/torch-2.6.0+cu128.nv-cp310-cp310-win_amd64.whl
# https://huggingface.co/w-e-w/torch-2.6.0-cu128.nv/resolve/main/torchvision-0.20.0a0+cu128.nv-cp310-cp310-win_amd64.whl

    
# Main requirements for insightface
insightface
scipy
opencv-python
requests
tqdm
numpy<2.0.0
onnxruntime-gpu>=1.17.0

# Optional/extra dependencies (for some features)
Pillow
matplotlib
scikit-learn
scikit-image
dlib
onnx
transformers>=4.30.0
accelerate>=0.20.0
protobuf>=3.20.0
nudenet
mediapipe

nsfw-detector
NudeNetClassifier

# Your existing NudeNet should already be installed

tensorflow>=2.8.0
tensorflow-hub>=0.12.0



